Update Engine.c Engine.h

+++ Engine.h
@@ -10,6 +10,7 @@
 
 #include <time.h>
 #include <stdint.h>
+#include <stdlib.h>
 #include "Moves.h"
 
 
@@ -21,7 +22,7 @@ int getScore(const struct gameState* gs);
 int miniMax(const struct gameState* gs, int depth, int alpha, int beta, bool playerColor);
 
 //acts as a wrapper for miniMax, returning bestMove, based off depth <--- determines difficulty
-uint16_t findBestMove(struct gameState* gs, int depth);
+uint32_t findBestMove(struct gameState* gs, int depth);
 
 //calls findBestMove (based off difficulty level), and makes the move
 void movePiece_Computer(struct gameState* gs, int difficulty);
@@ -30,8 +31,8 @@ void movePiece_Computer(struct gameState* gs, int difficulty);
 //returns a weight, weights are higher when attacker is low value
 //and captured piece is high
 //think: pawn captures queen
-int MVV_LVA(const struct gameState* gs, uint16_t move);
+int MVV_LVA(const struct gameState* gs, uint32_t move);
 //insertion sorts by weight (from MVV_LVA)
-void preSort(const struct gameState* gs, uint16_t* moves, int moveCount);
+void preSort(const struct gameState* gs, uint32_t* moves, int moveCount);
 
 #endif
